# static/public/openapi.yaml
openapi: 3.1.0
info:
  title: CanalGov API
  version: 0.1.0
  description: >
    Documentacao inicial da API do CanalGov para os fluxos principais de
    central de chamados. Esta base ainda nao cobre 100% das rotas do backend.
servers:
  - url: http://localhost:9000
    description: Ambiente local
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Autenticacao]
      summary: Autentica um agente
      operationId: loginAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sessao criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
...
  /api/v1/agents/me:
    get:
      tags: [Agentes]
      summary: Retorna o agente autenticado
      operationId: getCurrentAgent
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Dados do agente atual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
    put:
      tags: [Agentes]
      summary: Atualiza o perfil do agente autenticado
      operationId: updateCurrentAgent
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
  /api/v1/agents/compact:
    get:
      tags: [Agentes]
      summary: Lista agentes em formato compacto
      operationId: getAgentsCompact
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista compacta de agentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompactEntity'
  /api/v1/teams/compact:
    get:
      tags: [Equipes]
      summary: Lista equipes em formato compacto
      operationId: getTeamsCompact
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista compacta de equipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompactEntity'
  /api/v1/inboxes:
    get:
      tags: [Inboxes]
      summary: Lista inboxes
      operationId: getInboxes
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista de inboxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Inbox'
  /api/v1/statuses:
    get:
      tags: [Taxonomias]
      summary: Lista status de conversa
      operationId: getStatuses
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista de status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompactEntity'
  /api/v1/priorities:
    get:
      tags: [Taxonomias]
      summary: Lista prioridades
      operationId: getPriorities
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista de prioridades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompactEntity'
  /api/v1/tags:
    get:
      tags: [Taxonomias]
      summary: Lista tags
      operationId: getTags
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista de tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompactEntity'
    post:
      tags: [Taxonomias]
      summary: Cria uma tag
      operationId: createTag
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameOnlyRequest'
      responses:
        '200':
          description: Tag criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompactEntity'
  /api/v1/contacts:
    get:
      tags: [Contatos]
      summary: Lista contatos
      operationId: getContacts
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/OrderBy'
      responses:
        '200':
          description: Lista paginada de contatos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
  /api/v1/contacts/search:
    get:
      tags: [Contatos]
      summary: Busca contatos
      operationId: searchContacts
      security:
        - sessionAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resultado da busca
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
  /api/v1/contacts/{id}:
    get:
      tags: [Contatos]
      summary: Retorna um contato
      operationId: getContact
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Contato encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
    put:
      tags: [Contatos]
      summary: Atualiza um contato
      operationId: updateContact
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactRequest'
      responses:
        '200':
          description: Contato atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
  /api/v1/contacts/{id}/notes:
    get:
      tags: [Contatos]
      summary: Lista notas de um contato
      operationId: getContactNotes
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Lista de notas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactNote'
    post:
      tags: [Contatos]
      summary: Cria nota em um contato
      operationId: createContactNote
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactNoteRequest'
      responses:
        '200':
          description: Nota criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactNote'
  /api/v1/contacts/{id}/notes/{note_id}:
    delete:
      tags: [Contatos]
      summary: Remove nota de um contato
      operationId: deleteContactNote
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
        - name: note_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Nota removida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
  /api/v1/conversations/all:
    get:
      tags: [Conversas]
      summary: Lista todas as conversas visiveis
      operationId: getAllConversations
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Lista de conversas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
  /api/v1/conversations/unassigned:
    get:
      tags: [Conversas]
      summary: Lista conversas sem agente responsavel
      operationId: getUnassignedConversations
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista de conversas sem atribuicao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
  /api/v1/conversations/assigned:
    get:
      tags: [Conversas]
      summary: Lista conversas atribuidas ao agente
      operationId: getAssignedConversations
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista de conversas atribuidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
  /api/v1/conversations/mentioned:
    get:
      tags: [Conversas]
      summary: Lista conversas com mencoes ao agente
      operationId: getMentionedConversations
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Lista de conversas mencionadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
  /api/v1/conversations/search:
    get:
      tags: [Conversas]
      summary: Busca conversas
      operationId: searchConversations
      security:
        - sessionAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resultado da busca
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
  /api/v1/conversations:
    post:
      tags: [Conversas]
      summary: Cria uma nova conversa
      operationId: createConversation
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '200':
          description: Conversa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: canalgov_session
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    CompactEntity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UpdateAgentRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
          nullable: true
    Agent:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        availability_status:
          type: string
        roles:
          type: array
          items:
            type: string
        teams:
          type: array
          items:
            $ref: '#/components/schemas/CompactEntity'
    Inbox:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        channel:
          type: string
        enabled:
          type: boolean
        from:
          type: string
          nullable: true
    NameOnlyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
...